refactor(utils): `relayEvents` has moved to a shared utils file
