+++ lib/topologies/replset.js
@@ -16,7 +16,7 @@ const Interval = require('./shared').Interval;
 const createClientInfo = require('./shared').createClientInfo;
 const SessionMixins = require('./shared').SessionMixins;
 const isRetryableWritesSupported = require('./shared').isRetryableWritesSupported;
-const relayEvents = require('./shared').relayEvents;
+const relayEvents = require('../utils').relayEvents;
 const isRetryableError = require('../error').isRetryableError;
 
 const defaultAuthProviders = require('../auth/defaultAuthProviders').defaultAuthProviders;
